requirements.txt
crewai==0.86.0
google-generativeai==0.8.3
pillow==10.4.0
python-dotenv==1.0.1
